RunDataFileLocator is now a concrete class.
